modified:   applications/CustomKeymap/Actions/Actions.cpp 	modified:   applications/Drambo/Drambo.cpp 	modified:   applications/Drambo/Drambo.h 	modified:   applications/Matrix/MatrixBoot/MatrixBoot.cpp 	modified:   applications/Matrix/MatrixBoot/MatrixBoot.h 	modified:   devices/Device.h 	modified:   devices/JoyMatrix/Drivers/AnalogInput.cpp 	modified:   devices/JoyMatrix/Drivers/Dial.cpp 	modified:   devices/JoyMatrix/Drivers/Encoder.h 	modified:   os/MatrixOS.h 	modified:   os/framework/ComponentVariable.h 	modified:   os/framework/Framework.h 	modified:   os/framework/KeyEvent.h 	new file:   os/framework/PSRAM.h 	modified:   os/framework/Timer.cpp 	modified:   os/system/ComponentSetting.cpp 	modified:   os/system/FATFS.cpp 	modified:   os/system/KnobCenter.cpp 	modified:   os/system/KnobCenter.h 	modified:   os/system/MIDIAPP/Arpeggiator.cpp 	modified:   os/system/MIDIAPP/Arpeggiator.h 	modified:   os/system/MIDIAPP/ArpeggiatorUI.h 	renamed:    os/system/MIDIAPP/Chord.cpp -> os/system/MIDIAPP/Chorder.cpp 	renamed:    os/system/MIDIAPP/Chord.h -> os/system/MIDIAPP/Chorder.h 	renamed:    os/system/MIDIAPP/ChordUI.h -> os/system/MIDIAPP/ChorderUI.h 	new file:   os/system/MIDIAPP/DataManager.cpp 	modified:   os/system/MIDIAPP/KeyPadInput.cpp 	modified:   os/system/MIDIAPP/MidiAppUI.h 	modified:   os/system/MIDIAPP/MidiCenter.cpp 	modified:   os/system/MIDIAPP/MidiCenter.h 	modified:   os/system/MIDIAPP/MidiRouter.cpp 	modified:   os/system/MIDIAPP/Node.h 	modified:   os/system/MIDIAPP/NodeUI.h 	new file:   os/system/MIDIAPP/SeqData.h 	deleted:    os/system/MIDIAPP/SeqManager.cpp 	modified:   os/system/MIDIAPP/Sequencer.cpp 	new file:   os/system/MIDIAPP/Sequencer.h 	new file:   os/system/MIDIAPP/SequencerUI.h 	deleted:    os/system/MIDIAPP/Sequncer.h 	deleted:    os/system/MIDIAPP/SequncerUI.h 	modified:   os/system/Parameters.h 	modified:   os/ui/UI.cpp 	modified:   os/ui/component/ChannelButton.h 	modified:   os/ui/component/KnobButton.h 	modified:   os/ui/component/MixerFader.h 	modified:   os/ui/component/MultiPad_Drum.cpp 	modified:   os/ui/component/MultiPad_Note.cpp 	modified:   os/ui/component/MultiPad_Piano.cpp 	modified:   os/ui/component/UI4pxKnobVar.h

+++ applications/Matrix/MatrixBoot/MatrixBoot.h
@@ -16,7 +16,7 @@ class MatrixBoot : public BootAnimation {
   Timer timer;
 
   Point origin = Point(7, 1);
-  uint8_t counter;
+  uint16_t counter;
 
   uint8_t boot_phase;
   uint32_t boot_phase_1_tick_time = 0;
